newComms antenna patches
Actually I decided to add single config files for easier maintenance.
Please provide you edits and additions.

The dedicated patches come first, using BEFORE[zzzRealismOverhaul]
- CoatlAerospace
- CommNetAntennasExtension
- CxAerospace
- DMagicOrbitalScience
- JX2Antenna
- KerbalGPS
- MechJeb2
- NAU-ORIGAMI
- NearFutureConstruction
- ReStockPlus
- RN-Skylab
- RN-Soviet_Probes
- RN-US_Probes
- ROCapsules
- RP-1
- Stock_RO
- Taerobee
- TokamakIndustries
- UniversalStorage2
- VenStockRevamp

As you can see, the list is far from finished. (These are only the parts I got in my install)
- Please provide your own patches in the same manner, so using a dedicated file for easy maintenance

The generic patches come after the dedicated patches, using FOR[zzzRealismOverhaul] for the main part and AFTER[zzzRealismOverhaul] for the finalization part.

They provide some simple formulas to use the antennaPower of the ModuleDataTransmitter as a calculation base and in some cases even the mass of the part.
This is to avoid all generic patched antennas having the same values.